Move templateOnly component helper to core
