+++ core/tests/helpers/templates.ts
@@ -1,0 +1,11 @@
+export function templateOnlyComponentTemplate(template: string): string {
+  return `import { setComponentTemplate } from '@ember/component';
+  import { precompileTemplate } from '@ember/template-compilation';
+  import templateOnlyComponent from '@ember/component/template-only';
+  export default setComponentTemplate(
+    precompileTemplate('${template}', {
+      strictMode: true,
+    }),
+    templateOnlyComponent()
+  );`;
+}
