+++ demo-cards/routes/schema.js
@@ -1,11 +1,11 @@
 export default class Routes {
   routeTo(path) {
     if (path === 'welcome') {
-      return 'http://demo.com/cards/welcome';
+      return 'https://demo.com/cards/welcome';
     }
 
     if (path === 'about') {
-      return 'http://demo.com/cards/about';
+      return 'https://demo.com/cards/about';
     }
   }
 }
